adding create trade layout
