+++ app/src/main/java/com/pylons/loud/fragments/PylonCentralScreen/CreateTradeFragment.kt
@@ -1,0 +1,52 @@
+package com.pylons.loud.fragments.PylonCentralScreen
+
+import android.content.Context
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import com.pylons.loud.R
+import kotlinx.android.synthetic.main.fragment_create_trade.*
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class CreateTradeFragment : Fragment() {
+    private var listener: OnFragmentInteractionListener? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_create_trade, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        button_create_trade.setOnClickListener {
+            listener?.onCreateTrade()
+        }
+    }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        if (context is OnFragmentInteractionListener) {
+            listener = context
+        } else {
+            throw RuntimeException(context.toString() + " must implement OnListFragmentInteractionListener")
+        }
+    }
+
+    override fun onDetach() {
+        super.onDetach()
+        listener = null
+    }
+
+    interface OnFragmentInteractionListener {
+        fun onCreateTrade()
+    }
+
+}
