+++ app/src/main/res/layout/fragment_pylon_central_trade.xml
@@ -6,51 +6,13 @@
     android:layout_height="match_parent"
     tools:context=".fragments.PylonCentralScreen.PylonCentralTradeFragment">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+
+    <fragment
+        android:id="@+id/nav_host_fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <LinearLayout
-            android:id="@+id/linearLayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/text_create_trade"
-                android:layout_margin="@dimen/text_margin"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/create_trade" />
-
-            <TextView
-                android:id="@+id/text_my_trades"
-                android:layout_margin="@dimen/text_margin"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/my_trades" />
-
-            <TextView
-                android:id="@+id/text_market"
-                android:layout_margin="@dimen/text_margin"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/market_trades" />
-        </LinearLayout>
-
-
-        <fragment
-            android:id="@+id/fragment_trade_list"
-            android:name="com.pylons.loud.fragments.trade.TradeFragment"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/linearLayout" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph_pylon_central_trade" />
 
 </FrameLayout>
